Update to be chunked database exports
- Added date start and date end
- Fix name of spec to sspy
- Fixup typos
- Add Win10 build instructions
- Chunk database export in to sets of 1024 items

+++ sspy/mysqldump.py
@@ -265,35 +265,33 @@ class Exporter(object):
         for table in tables:
             reset_line()
             print_message("\r[%d/%d] - %s" % (index, tables_l, table), False)
+            count = self.fetch_one("SELECT COUNT(`ID`) FROM `%s`" % table)
 
-            columns = self.fetch_single(
-                "SELECT `column_name` FROM `information_schema`.`columns` "
-                "WHERE `table_schema` = '%s' AND `table_name` = '%s'" %
-                (self.database, table)
-            )
-            columns_s = "`, `".join(columns)
-            data = self.fetch_all(
-                "SELECT `%s` FROM `%s`" % (columns_s, table)
-            )
-
-            self._write_file(file, "DELETE FROM `%s`;\n" % table)
-            if len(data):
+            if count[0] > 0:
+                i = 0
+                columns = self.fetch_single(
+                    "SELECT `column_name` FROM `information_schema`.`columns` "
+                    "WHERE `table_schema` = '%s' AND `table_name` = '%s'" %
+                    (self.database, table)
+                )
+                columns_s = "`, `".join(columns)
+                self._write_file(file, "DELETE FROM `%s`;\n" % table)
                 self._write_file(file, "/*!40000 ALTER TABLE `%s` DISABLE KEYS */;\n" % table)
                 self._write_file(file, "/*!40000 ALTER TABLE `%s` ENABLE KEYS */;\n\n" % table)
-                # Split in to chunks of 100 items per chunk to not crash the database
-                chunks = [data[x:x + 1024] for x in range(0, len(data), 1024)]
-                for chunk in chunks:
-                    self._write_file(file, "INSERT INTO `%s` (`%s`) VALUES " % (table, columns_s))
-                    self.dump_data(file, chunk)
+                while i <= count[0]:
+                    data = self.fetch_all(
+                        "SELECT `%s` FROM `%s` LIMIT 1024 OFFSET %s" % (columns_s, table, int(i))
+                    )
+                    self._write_file(file, "INSERT INTO `%s` (`%s`) VALUES \n" % (table, columns_s))
+                    self.dump_data(file, data)
+                    i = i + 1024
             index += 1
 
         # Reset the SQL Modes to what it's supposed to be :)
-        self._write_file(file, "/*!40101 SET SQL_MODE=IFNULL(@OLD_SQL_MODE, '') */;\n")
-        self._write_file(
-            file,
-            "/*!40014 SET FOREIGN_KEY_CHECKS=IF(@OLD_FOREIGN_KEY_CHECKS IS NULL, 1, @OLD_FOREIGN_KEY_CHECKS) */;\n"
-        )
-        self._write_file(file, "/*!40101 SET CHARACTER_SET_CLIENT=@OLD_CHARACTER_SET_CLIENT */;")
+        self._write_file(file, "/*!40101 SET SQL_MODE=IFNULL(@OLD_SQL_MODE, '') */;\n"
+                               "/*!40014 SET FOREIGN_KEY_CHECKS=IF(@OLD_FOREIGN_KEY_CHECKS IS NULL, 1,"
+                               " @OLD_FOREIGN_KEY_CHECKS) */;\n"
+                               "/*!40101 SET CHARACTER_SET_CLIENT=@OLD_CHARACTER_SET_CLIENT */;")
 
         file.close()
         final = time.time()
